New configuration for controlling post selection.
This commit adds a number of additional capabilities:

1. The ability to control inclusion of site pages.
2. Collection-level inclusion controls.
3. Exclusion regexs so you can selectively remove pages from shortening.

This commit also changes the shortener cache to use URLs as the page
IDs, as URLs are available for all page types.  The downside is that if
you change the URL for a page (e.g. the permalink) the shortened URL
changes as well.  But in the end that's probably correct, anyway...